fix: refactor repository to fit new name
fixes #14
