+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@open-rpc/introspector",
+  "name": "@open-rpc/logs-react",
   "version": "0.0.0-development",
   "description": "",
   "main": "package/index.js",
